create USBClass as separate module
